Update BetaBuild.py
Added functionality for Winn Dixie as well as multiple other Kroger brand of companies stores. 

Note: Winn Dixie has issues with coupon clicking even manually. Seems to only impact coupons that are within 1 day of expiry. Not making alterations for those specific coupons at this time, as the remainder of the program works and even manually these items are not addable as digital coupons.